feat: implement RegisterHandler for user registration and validation
