+++ infrastructure/entry-points/reactive-web/src/main/java/co/com/crediya/api/user/register/dto/mapper/UserMapperDto.java
@@ -7,8 +7,5 @@ import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
 public interface UserMapperDto {
-
-    User toDomain(CreateUserRequestDto createUserRequestDto);
-
     CreateUserResponseDto toResponse(User user);
 }
